fixed bug with generating of files move generation to service, allowed count in background refactored project structures added test flavour

+++ app/src/main/java/com/ideas/sportscounter/timer/VibratorHelper.java
@@ -24,7 +24,7 @@ public class VibratorHelper {
         enabled = preferences.isVibroEnabled();
     }
 
-    public void vibrate(float seconds) {
+    public void vibrate(int seconds) {
         if (!enabled) {
             return;
         }
